pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-organizer"
version = "1.0.0"
authors = [
    {name = "Deep Organizer"},
]
description = "AI-powered file organization tool that intelligently organizes files based on content analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
keywords = [
    "ai",
    "file-organization",
    "automation",
    "cli",
    "productivity", 
    "file-management",
    "langchain",
    "openai",
    "gpt"
]
dependencies = [
    "langgraph",
    "langchain[anthropic]",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

[project.scripts]
deep-organizer = "deep_organizer.cli:main"
deep_organizer = "deep_organizer.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/deep-organizer"
"Bug Reports" = "https://github.com/yourusername/deep-organizer/issues"
"Source" = "https://github.com/yourusername/deep-organizer"
"Documentation" = "https://github.com/yourusername/deep-organizer#readme"

[tool.setuptools]
packages = ["deep_organizer"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=deep_organizer",
    "--cov-report=term-missing",
    "--cov-report=html",
]